ml-model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0
albumentations>=1.3.0
scikit-image>=0.20.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
boto3>=1.28.0
botocore>=1.31.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
requests>=2.31.0
geopy>=2.3.0
reportlab>=4.0.0
jinja2>=3.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
psutil>=5.9.0